split single `senum.rs` file into enums module with multiple files.
