+++ src/blocking/client.rs
@@ -34,7 +34,7 @@ use crate::model::show::{
 };
 use crate::model::track::{FullTrack, FullTracks, SavedTrack, SimplifiedTrack};
 use crate::model::user::{PrivateUser, PublicUser};
-use crate::senum::{
+use crate::enums::{
     AdditionalType, AlbumType, Country, IncludeExternal, RepeatState, SearchType, TimeRange, Type,
 };
