+++ src/model/playlist.rs
@@ -8,7 +8,7 @@ use super::image::Image;
 use super::page::Page;
 use super::track::FullTrack;
 use super::user::PublicUser;
-use crate::senum::Type;
+use crate::enums::Type;
 
 ///[Playlist object simplified](https://developer.spotify.com/web-api/object-model/#playlist-object-simplified)
 #[derive(Clone, Debug, Serialize, Deserialize)]
